refactor(usage): split goUsageTracker into src/usage/ modules
Split the 1510-line goUsageTracker.ts god file into the usage domain:
- usage/tracker.ts — GoUsageTracker class + types + time-window helpers
- usage/history.ts — OpenCode CLI SQLite history read/aggregation (pure)
- usage/pricing.ts — bundled pricing snapshot + estimateCost (pure)
- usage/formatting.ts — status-bar / quick-pick formatting

Move usage.ts / usageProfile.ts / goUsageSync.ts into src/usage/ and keep
goUsageTracker.ts as a thin barrel re-exporting the historical public API.
Behavior-preserving; all importers + tests updated.
